refactor: add hooks folder and move to this time hook
